fixed create tweet

+++ src/main/java/com/twitter/entity/Tweet.java
@@ -7,7 +7,7 @@ public class Tweet {
     private int id;
     private String content;
     private String createdTime;
-    private long userId;
-    private long commentCount;
-    private long likeCount;
+    private int userId;
+    private int commentCount;
+    private int likeCount;
 }
